% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward-look.R
\name{forward_look}
\alias{forward_look}
\title{Get research and statistics announcements from GOV.UK}
\usage{
forward_look(
  organisation = "UK Health Security Agency",
  type = "upcoming statistics",
  start_date = NULL,
  end_date = NULL,
  drop_weekly = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{organisation}{UK Government department, agency or public body.}

\item{type}{Type of announcement: all research and statistics, published statistics, upcoming statistics, cancelled statistics or research.}

\item{start_date}{Currently not used.}

\item{end_date}{Currently not used.}

\item{drop_weekly}{If \code{TRUE}, announcements for weekly publications are removed.}

\item{verbose}{If \code{TRUE}, the default, update message(s) are returned.}
}
\value{
A tibble.
}
\description{
Get research and statistics announcements from GOV.UK
}
\examples{
forward_look("Department of Health and Social Care")
}
