% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardlook-package.R
\docType{package}
\name{forwardlook-package}
\alias{forwardlook}
\alias{forwardlook-package}
\title{forwardlook: Get research and statistics announcements from GOV.UK}
\description{
`forwardlookR` contains functions to automatically scrape and clean research and statistics announcements from GOV.UK.
}
\author{
\strong{Maintainer}: Si Maxwell \email{simon.maxwell@ukhsa.gov.uk}

}
\keyword{internal}
