% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-announcement-info.R
\name{get_announcement_info}
\alias{get_announcement_info}
\title{Get GOV.UK research and statistics announcement information}
\usage{
get_announcement_info(url, verbose = TRUE)
}
\arguments{
\item{url}{URL of a research or statistics announcement.}

\item{verbose}{If \code{TRUE}, the default, update message(s) are returned.}
}
\value{
A list.
}
\description{
Get GOV.UK research and statistics announcement information
}
\examples{
url <- "a statistics announcement url"
get_announcement_info(url)
}
