% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-announcement-url.R
\name{get_announcement_url}
\alias{get_announcement_url}
\title{Get GOV.UK research and statistics announcement URLs}
\usage{
get_announcement_url(
  organisation = "UK Health Security Agency",
  type = "upcoming statistics",
  verbose = TRUE
)
}
\arguments{
\item{organisation}{UK Government department, agency or public body.}

\item{type}{Type of announcement: all research and statistics, published statistics, upcoming statistics, cancelled statistics or research.}

\item{verbose}{If \code{TRUE}, the default, update message(s) are returned.}
}
\value{
A character vector of announcement URLs.
}
\description{
Get GOV.UK research and statistics announcement URLs
}
\examples{
get_announcement_url()
}
