% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-org-info.R
\name{get_org_info}
\alias{get_org_info}
\title{Get list of organisations in UK Government}
\usage{
get_org_info()
}
\value{
A tibble.
}
\description{
Get list of organisations in UK Government
}
\examples{
get_org_info()
}
