% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{govuk_orgs}
\alias{govuk_orgs}
\title{List of UK Government organisations from GOV.UK API}
\format{
A tibble with 1223 rows and 3 columns:
\describe{
\item{title}{Name of UK Government department, agency or public body}
\item{url}{URL of organisation homepage on GOV.UK}
\item{slug}{The organisation identifier used in the GOV.UK API}
}
}
\usage{
govuk_orgs
}
\description{
A data set containing the organisation name and homepage URL of over 1,200 departments, agencies and public bodies as listed on GOV.UK.
}
\keyword{datasets}
