% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-org-info.R
\name{update_org_info}
\alias{update_org_info}
\title{Gather and save list of organisations in UK Government}
\usage{
update_org_info()
}
\value{
Nothing. List of organisations saved in data/
}
\description{
Gather and save list of organisations in UK Government
}
\examples{
update_org_info()
}
